backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow>=10.4.0

# Celery and Redis
celery==5.3.4
redis==5.0.1

# Cloudinary for image storage
cloudinary==1.38.0
django-cloudinary-storage==0.3.0

# Playwright for browser automation
playwright==1.40.0

# HTTP requests
requests==2.31.0

# OpenAI for AI-powered features
openai==1.12.0

# Jira for bug tracking integration
jira>=3.5.0